Collapse CreateIndexProcedure tests into one
